[FIX] diel tomato model + Both validations + DFA + PCA & T-SNE
